Make FileDialog look native
As a side effect, use a Loader to load the
main updater window. Previously, we were creating
a component and then creating an object. This was
causing problems for an unknown reason where closing
the FileDialog would close the updater. Using an updater
fixes this problem (probably handles the parent heirarchy
correctly...).