game homepath: migrate unix legacy dot dir to xdg data dir
The engine knows how to do it but if the updater is started before the engine
(and that's the expected behavior), the updater will create the directory
and the engine will not do the migration itself since both legacy path and
new one will exist. That's why the updater move the game assets the sooner it
can.