Windows: add application manifest declaring requested privilege level.
The point is to disable the installer detection heuristics. When
installer detection is triggered, you get an annoying pop-up saying
"this program may not have installed correctly", if you decide to close
the program without completing an installation. The installer already
always ran as administrator to to installer detection, but by having the
manifest the annoying messages can be avoided.