On Windows, launch the game as non-administrator
This was achieved using a deeply magical code sample from Microsoft
which gets the Explorer process for the current user to start the
process. This code is used in place of QProcess.

Annoyingly, the function they used requires the executable and arguments
to be specified separately, so I had to manually tokenize the first arg
of the command line. At least Windows filenames can't contain quote or
backlash, so I can ignore such cases.