Linux Docker: Build Qt 5.14
Now it is built in the Dockerfile instead of using a prebuilt image,
like the Windows one.

Fortunately I was able to recover the commands used to build the old
image via `docker history` and reuse some of the knowledge.

I diffed the output of ldd before and after the change to make sure the
build is as "static" as before. The result is that libXi.so.6 was
removed and librt.so.1 and libxkbcommon.so.0 are added.