After starting game, exit by closing all windows
It's a default behavior of Qt to exit when there are no windows. In the
case of a failure to start the process, the program stays alive until
the error QMessageBox is closed.

Now the updater will exit as soon as Unanquished starts on Linux/Mac (it
already does on Windows).