Rework handling of currentVersion and installPath settings
Persist the installPath setting only after the user has started the
update and the path has been validated. So, for example, if you set the
path and then immediately close the application, the setting will not
change.

currentVersion (i.e. installed version):
- Set to "" if a download is started with a different path than before
- Set to "" before extracting the platform-specific zip
- Set to the current version after extracting the platform-specific zip

The goal is to have currentVersion be non-empty if and only if there is
currently a usable installation in installPath.