Windows: use global instead of per-user registry keys
To make the launcher runnable as non-admin, we need to be able to read
the same registry keys whether elevated or not. The admin user can be
different than the non-admin user after UAC elevation which would change
the registry values we get. For the moment I am not planning to support
installing without UAC elevation, so just use global keys.