Windows: Let updater run as non-administrator
Change the executable manifest to not request elevation.

Now, administrator is required only when doing an updater update or
opening the main window. When it is time to perform one of these
actions and the updater is not running as administrator, it tries to
relaunch using UAC elevation, passing the state in command-line flags.

If the user clicks the settings button but elevation fails, the program
exits.
If an updater or game update is needed according to the version check
and elevation fails, the updater falls back to launching the currently
installed game version, if there is one.