Produce release artifacts in Docker builds
Produce correctly named Windows and Linux zips and Windows binary in the
Docker bulids, which can be directly used in a release.

For the Linux binary, I chose the name 'unvanquished-updater' instead of
'updater2' which was used last time.

I decided 'zip -9' was not worth it since it took a few seconds and
changed the file size only 0.3%.