Set installed version from download, not versions.json
When saving the installed version ("currentVersion") to the settings,
base the version on the direcory name in the torrent
(unvanquished_0.54.1/...) rather than what versions.json says. This will
be more robust against the versions request failing (previously the
version would be set to "unknown" in that case) or versions.json being
out of sync with the download.

Now the versions.json file will only be used to decide to automatically
go to the game update screen