Connect to a URL passed on the command line
The updater now accepts a positional argument on the command line which
is the address of an Unvanquished server to connect to upon
starting the game. If the settings/installation window is entered, the
URL will be intentionally forgotten (if it took an hour to update, the
desire to connect to a particular server has likely passed).

For now we do just `daemon -connect <address>` when launching with a
URL, ignoring the custom command line. This behavior matches the Windows
and Linux unv:// protocol handlers But I want to support the custom
command line with the next Unvanquished release (which will have
-connect-trusted). So the URL is restricted to not have any shell
metacharacters to minimize the chances of injecting code into a custom
command line.