Conceal command line options for internal use
The --update-updater-to, --update-game, and --playnow options are
removed and replaced with a single option --internalcommand, which can
do any one of those three functions. This is done because
(a) The options are intended for the updater to use when restarting
itself, not for the user, and may not work as expected if used by the
user. For example if the user used --updater-update-to on Windows, it
would not perform the required admin elevation.
(b) The three options are intended to be mutually exclusive.