Toggle or stop download thread-safely
The download loop which downloads using libaria2 runs in a secondary
thread. The downloader's toggle and stop functions were called from the
main thread. The 'stop' case was probably harmless in practice, but the
toggle case manipulates the aria objects which are concurrently used by
the download thread in a dangerous way.

Fix this by defining toggle and stop as signals which are sent to the
downloader thread, and have the download loop manually check the
thread's event queue.