Dockerfiles: use fully qualified image names
This lets podman be used as a substitute for docker, and avoids package
squatter vulnerabilities when there are multiple registries.

I'm messing with the EOL Debian images since if I just change it to
Bullseye, icons are broken in the update. And if we use the old name of
the Buster image, APT doesn't work.