Close splash screen instead of hiding it
With Qt 5, this fixes a bug where (on Linux at least) the entire program
closes if you click the X in the file dialog. 

On Qt 6 things are much more broken without this commit. If the splash
screen is just hidden instead of closed, it keeps the application alive
indefinitely. So we must close the splash screen instead of hiding it to
use the automatic exiting feature. But with the way we were creating the
windows before, the other windows were descendants of the splash screen
for memory management purposes and closing the splash screen would
remove them. So I added a new global object to be the QObject parent for
other top-level windows.